+++ src/deployconfig.jl
@@ -821,6 +821,58 @@ authentication_method(::Buildkite) = Documenter.SSH
 
 documenter_key(::Buildkite) = ENV["DOCUMENTER_KEY"]
 
+"""
+    SimpleSSHConfig <: DeployConfig
+
+Implementation of deploy config for deploying directly via SSH.  This is particularly useful in cases
+in which there is a private repository and it's impractical or impossible to set up GitHub Actions or
+Travis for fully automated deployment.
+
+The constructor optionally accepts a path to the SSH key, so that it is possible to use multiple distinct keys.
+If the key location is not passed or is an empty string, it will check for the `DOCUMENTER_KEY_PATH` environment
+variable for a path to an SSH key.  If `DOCUMENTER_KEY_PATH` is not set, it will check for `DOCUMENTER_KEY`
+as in the Travis and GitHub Actions configs.  Failing this, it will check for a key at `\$HOME/.ssh/id_rsa`.
+
+!!! warning "Tagged versions"
+
+    `SimpleSSHConfig` does not currently support pushing documentation for tagged versions, and
+    will always deploy the documentation into development version directory in the remote
+    repository.
+"""
+struct SimpleSSHConfig <: DeployConfig
+    ssh_key_location::String
+end
+
+SimpleSSHConfig() = SimpleSSHConfig("")
+
+_read_ssh_key(fname) = base64encode(open(read, fname))
+
+function deploy_folder(::SimpleSSHConfig; repo, devbranch, push_preview, devurl,
+                       branch="gh-pages", kwargs...)
+    DeployDecision(all_ok=true, repo=repo, branch=branch, subfolder=devurl)
+end
+authentication_method(::SimpleSSHConfig) = SSH
+function documenter_key(cfg::SimpleSSHConfig)
+    k = cfg.ssh_key_location
+    if isempty(k)
+        k1 = get(ENV, "DOCUMENTER_KEY_PATH", nothing)
+        k2 = get(ENV, "DOCUMENTER_KEY", nothing)
+        if k1 !== nothing
+            @info "SimpleSSHConfig: Using key location from `DOCUMENTER_KEY_PATH`"
+            _read_ssh_key(expanduser(k1))
+        elseif k2 !== nothing
+            @info "SimpleSSHConfig: Using key from `DOCUMENTER_KEY`"
+            k2
+        else
+            @info "SimpleSSHConfig: Using key at `~/.ssh/id_rsa`"
+            _read_ssh_key(joinpath(ENV["HOME"],".ssh","id_rsa"))
+        end
+    else
+        base64encode(open(read, k))
+    end
+end
+
+
 ##################
 # Auto-detection #
 ##################
@@ -829,6 +881,8 @@ function auto_detect_deploy_system()
         return Travis()
     elseif haskey(ENV, "GITHUB_REPOSITORY")
         return GitHubActions()
+    elseif haskey(ENV, "DOCUMENTER_KEY_PATH")
+        return SimpleSSHConfig()
     elseif haskey(ENV, "GITLAB_CI")
         return GitLab()
     elseif haskey(ENV, "BUILDKITE")